adding rejected path to mod controller
moved rejected function from userController to moderatorController
